Adds delete function
Delete now works on current project.  Also added .empty() when running
`GET` request so fields are not appended with multiple copies of the
update or delete.  Since the `GET` request is run on each delete,
the view updates to reflect the deletion.

Next step is to add logic to prevent user from creating a new project
while they still have a project in the works.

It'll be something like:

if data.projects.length === 0, show create projects, else hide it.