Hides create button, adds `GET` to signInSuccess
Adds logic to app which counts how many Projects there are.  Since
there can only be one, the logic is set to hide the Create button
if there's a Project.  Once the project is deleted, the button
reappears.

Also, adds `GET` (getProject) to signInSuccess.  This triggers the
new logic function which is located in getProjectSuccess:

if (data.projects.length === 1)
    $('#new-project-modal-button').addClass('hidden')
  else {
    $('#new-project-modal-button').removeClass('hidden')
  }