Fixes 'GET' on sign-in
'GET' on sign in did not work properly before.  Now, GET fires and
populates creation-content if there is a Project.
If there is not a Project, it removes the hidden class from the
createProject button to allow the user to create a Project.

Next step is to fix the bug where the createPRoject button remains after
Project has been created.  This is a bug because the user can over-
write their first Project.