Fixes bug that allowed Create Project button to persist
Users were able to overwrite Projects since Create Button persisted

Fixes that bug with logic that GETs list at sign-in.  If there are no
Projects, the button become visible.  If there is a Project, the
button remains hidden.

Also, fixes bug that just showed users "Null" when they did not enter
anything into Create.  They now see a message that lets them know
if they did not enter a Title or a Description or both.